args: command-line arguments and help text
